requirements-ml.txt
# Optional: DistilRoBERTa text emotion + ViT face emotion (FER-style labels).
# Install after base requirements: pip install -r requirements.txt -r requirements-ml.txt
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.25.0
Pillow>=10.0.0
safetensors>=0.4.0

# Optional AWS mood backend
# boto3>=1.34.0
